convert uint64 to int64 for BSON no support uint64

+++ modules/gov/vote.go
@@ -6,7 +6,7 @@ import (
 
 // MsgVote
 type DocTxMsgVote struct {
-	ProposalID uint64 `bson:"proposal_id"` // ID of the proposal
+	ProposalID int64  `bson:"proposal_id"` // ID of the proposal
 	Voter      string `bson:"voter"`       //  address of the voter
 	Option     int32  `bson:"option"`      //  option from OptionSet chosen by the voter
 }
@@ -19,7 +19,7 @@ func (doctx *DocTxMsgVote) BuildMsg(txMsg interface{}) {
 	msg := txMsg.(*MsgVote)
 	doctx.Voter = msg.Voter
 	doctx.Option = int32(msg.Option)
-	doctx.ProposalID = msg.ProposalId
+	doctx.ProposalID = int64(msg.ProposalId)
 }
 
 func (m *DocTxMsgVote) HandleTxMsg(v SdkMsg) MsgDocInfo {
